#558 likeボタン、bookmarkボタン、フォローボタンをログインせずに押すと、contentmissingが表示される…

+++ app/helpers/application_helper.rb
@@ -25,6 +25,11 @@ module ApplicationHelper
   end
 
   def create_follow_button(user_id)
+    if !user_signed_in?
+      return button_to "フォローする", user_relationships_path(user_id), id: "follow_button", class: "follow-button",
+                                                                   form: { data: { turbo_frame: "_top" } }
+    end
+
     if user_signed_in? && current_user.following?(user_id)
       button_to "フォロー中", user_relationships_path(user_id),
       method: :delete, data: { turbo_confirm: 'フォローをやめますか?' }, id: "follow_button", class: "follow-button"
